requirements.txt
antlr4-python3-runtime==4.9
Pillow==8.1.0
python-telegram-bot==13.1
